+++ frontend/src/pages/admin/AdminKomentarji.jsx
@@ -1,8 +1,10 @@
 import { useEffect, useState } from "react";
 import { adminFetch } from "./adminApi";
+import { useNavigate } from "react-router-dom";
 
 export default function AdminKomentarji() {
   const [kom, setKom] = useState([]);
+  const navigate = useNavigate();
 
   const load = async () => {
     const data = await adminFetch("/admin/komentarji");
@@ -10,6 +12,11 @@ export default function AdminKomentarji() {
   };
 
   const del = async (id) => {
+    const confirmDelete = window.confirm(
+      "Ali si prepričan, da želiš izbrisati komentar?",
+    );
+
+    if (!confirmDelete) return;
     await adminFetch(`/admin/komentarji/${id}`, { method: "DELETE" });
     load();
   };
@@ -23,9 +30,17 @@ export default function AdminKomentarji() {
       <h1 className="text-xl font-bold mb-4">Komentarji</h1>
 
       {kom.map((k) => (
-        <div key={k.id} className="bg-white p-3 border mb-2 flex justify-between">
-          <p>{k.vsebina}</p>
-          <button onClick={() => del(k.id)} className="text-red-500">
+        <div
+          key={k.id}
+          className="bg-white p-3 border mb-2 flex justify-between"
+        >
+          <div
+            onClick={() => navigate(`/objava/${k.objava_id}`)}
+            className="cursor-pointer hover:underline"
+          >
+            <p>{k.vsebina}</p>
+          </div>
+          <button onClick={() => del(k.id)} className="text-red-500 cursor-pointer hover:underline">
             Delete
           </button>
         </div>
